nonceはint32

+++ protoc-gen-js/eth/commonMessage_Eth_pb.js
@@ -766,7 +766,7 @@ proto.fairwaycorp.blockchainprotobuf.eth.Block.deserializeBinaryFromReader = fun
       msg.setRelayedBy(value);
       break;
     case 13:
-      var value = /** @type {number} */ (reader.readUint64());
+      var value = /** @type {number} */ (reader.readUint32());
       msg.setNonce(value);
       break;
     case 14:
@@ -924,7 +924,7 @@ proto.fairwaycorp.blockchainprotobuf.eth.Block.serializeBinaryToWriter = functio
   }
   f = message.getNonce();
   if (f !== 0) {
-    writer.writeUint64(
+    writer.writeUint32(
       13,
       f
     );
@@ -1183,7 +1183,7 @@ proto.fairwaycorp.blockchainprotobuf.eth.Block.prototype.setRelayedBy = function
 
 
 /**
- * optional uint64 nonce = 13;
+ * optional uint32 nonce = 13;
  * @return {number}
  */
 proto.fairwaycorp.blockchainprotobuf.eth.Block.prototype.getNonce = function() {
